Fix paths for Atom
