+++ problems/P002.java
