+++ problems/P003.java
