+++ problems/P006.java
