+++ problems/P010.java
