+++ problems/P012.java
